Add ControllerBaseTestCase
